Add InferenceGraph type
